Partition delete uses ducklake Add --to to partition delete move compact and delete into new maintenance.go
